util and data type

+++ model/car_file.go
@@ -4,10 +4,10 @@ type CarFile struct {
 	Id         int     `json:"id"`
 	TaskId     int     `json:"task_id"`
 	OriginName string  `json:"origin_name"`
-	StartEpoch int     `json:"start_epoch"`
+	StartEpoch int64   `json:"start_epoch"`
 	FileUrl    string  `json:"file_url"`
 	FileMd5    *string `json:"file_md5"`
-	FileSize   int     `json:"file_size"`
+	FileSize   int64   `json:"file_size"`
 	PayloadCid string  `json:"payload_cid"`
 	PieceCid   string  `json:"piece_cid"`
 	PinStatus  *string `json:"pin_status"`
